Version 1.2 > Limit bandwidth to 1200M > On record - check if file exists > Read the stdout and error and display on status line > Add the Recorder.java (.proto java version) to the git
Bugfix:
> Change nsumps to long in java and int64 in the proto file

+++ src/jrecorder/ManagmentClient.java
@@ -246,7 +246,7 @@ class ManagementClient extends WebSocketClient
 	}
 
 	public Boolean SendRecordCommand(double CenterFrequncy, double Rate, double Gain, double bw, String Filename,
-			int NumberOfSampels, String RecordExe)
+			long NumberOfSampels, String RecordExe)
 	{
 		RecordCommand s = RecordCommand.newBuilder().setFrequency(CenterFrequncy).setRate(Rate).setGain(Gain).setBw(bw)
 				.setFilename(Filename).setNumberOfSamples(NumberOfSampels).setApplicationExecute(RecordExe).build();
